Small cleanup + Fix memory base logged as 0
