+++ src/gui/GameUpdateWindow.cpp
@@ -34,8 +34,6 @@ std::string _GetTitleIdTypeStr(TitleId titleId)
 	return "Unknown";
 }
 
-bool IsRunningInWine();
-
 bool GameUpdateWindow::ParseUpdate(const fs::path& metaPath)
 {
 	m_title_info = TitleInfo(metaPath);
@@ -130,15 +128,11 @@ bool GameUpdateWindow::ParseUpdate(const fs::path& metaPath)
 		}
 	}
 
-	// checking size is buggy on Wine (on Steam Deck this would return values too small to install bigger updates) - we therefore skip this step
-	if(!IsRunningInWine())
+	const fs::space_info targetSpace = fs::space(ActiveSettings::GetMlcPath());
+	if (targetSpace.free <= m_required_size)
 	{
-		const fs::space_info targetSpace = fs::space(ActiveSettings::GetMlcPath());
-		if (targetSpace.free <= m_required_size)
-		{
-			auto string = wxStringFormat(_("Not enough space available.\nRequired: {0} MB\nAvailable: {1} MB"), L"%lld %lld", (m_required_size / 1024 / 1024), (targetSpace.free / 1024 / 1024));
-			throw std::runtime_error(string);
-		}
+		auto string = wxStringFormat(_("Not enough space available.\nRequired: {0} MB\nAvailable: {1} MB"), L"%lld %lld", (m_required_size / 1024 / 1024), (targetSpace.free / 1024 / 1024));
+		throw std::runtime_error(string);
 	}
 
 	return true;
